Added GUI and Tile Classes
Added new classes to the project. Tile is going to be used to create the board, and eventually we should be able to use it to move the tokens around the board. GUI is the main interface that players will interact with when they are actually playing the game. Right now the biggest issue with the GUI is scaling the game board properly so that other aspects of the GUI, such as the dice area and the button box, will always fit on screen. I also made a few changes to the BoardBuilder class in order to be able to send it to the new GUI class.