Prototype 1
The first prototype for our project. Has a working GUI, which is able can be closed by either clicking on the close button in the corner, or by going into File -> Exit from the menu. The only other button with functionality right now is the "Roll" button, which will change the images displayed in the right pane to reflect the player's roll. Player tokens to be added later on.